rename appReducer to userReducer; fix user state
