+++ client/src/components/AlcoCounter/alcoComponents/Display/Display.tsx
@@ -9,7 +9,7 @@ export const Display = () => {
     (state) => state.alcoReducer
   );
   const { currentTheme } = useAppSelector(
-    (state) => state.appReducer
+    (state) => state.userReducer
   );
   const { day, month, year } = alcoState.currentDate;
 
@@ -23,7 +23,7 @@ export const Display = () => {
     getTotalDrankData(alcoState);
 
   const { currentLang } = useAppSelector(
-    (state) => state.appReducer
+    (state) => state.userReducer
   );
 
   useEffect(() => {
